Release Version now available!
Version 1.0 has been in the making for the past month and is finally in a finished and stable state! We hope you find DMG Reader useful and look forward to any feedback you might have!

Release Version now includes
-Execution of files by double-clicking
This makes DMG-Reader the easiest program to view DMGs publicly available

-DMG files now have updated icon
With the new icon, MacOS users can feel even more at-home on Windows.

-Setup program
The inclusion of the new Setup program negates the need of any unnecessary steps the user previously needed to take in order to get DMG Reader running.

-Multiple ways of opening DMGs
With the inclusion of double-clicking to open DMGs, not only is it easier, but provides one extra way and efficient way to open these files.

-And much more
With extra design improvements, we feel users will be pleased to use this program going forward